fix: Use uint16_t for IoHomeChannel_t.c0 to prevent frequency overflow

+++ src/IoHome.h
@@ -71,7 +71,7 @@ struct NodeId {
 };
 
 struct IoHomeChannel_t {
-  uint8_t c0;
+  uint16_t c0;
   uint8_t c1;
 };
